Double precision math functions changed to single precision

+++ src/types.h
@@ -158,7 +158,7 @@ static inline vec3_t vec3_divf(vec3_t a, float f) {
 }
 
 static inline float vec3_len(vec3_t a) {
-	return sqrt(a.x * a.x + a.y * a.y + a.z * a.z);
+	return sqrtf(a.x * a.x + a.y * a.y + a.z * a.z);
 }
 
 static inline vec3_t vec3_cross(vec3_t a, vec3_t b) {
